logisticsnetwork/requirements.txt
Django==2.1.8
djangorestframework==3.9.0
psycopg2==2.7.6.1
factory-boy==2.11.1
networkx==2.3
